extract base class from image tile to inherit from for video tile
